Removed the proxy configuration in package.json as it was not working with s3. Made changes reflecting those in api.js
S3 was throwing up when it was trying to make a call through proxy, it was using its url, instead of the proxy's url.

I havent made any changes to package-json, but they got added. Need info from Ashish on why that happened

I tested the code both on local and on the dev server, it is being deployed as expected.